--- src/components/common/Card/index.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const Card = styled.div`
-  padding: 1rem;
-  background: #fff;
-  height: 100%;
-`;
