+++ src/components/common/Layout/Layout.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Global, Main } from './styles';
 import './fonts.css';
 
@@ -8,3 +9,7 @@ export const Layout = ({ children }) => (
     {children}
   </Main>
 );
+
+Layout.propTypes = {
+  children: PropTypes.array,
+};
