+++ src/components/common/SEO/index.js
@@ -1,0 +1,1 @@
+export { SEO } from './SEO';
