+++ src/components/common/index.js
@@ -1,8 +1,7 @@
-export * from './Layout';
-export * from './Container';
-export * from './Button';
-export * from './Card';
-export * from './Input';
-export * from './SEO';
-export * from './Divider';
-export * from './SectionHeader';
+export { Layout } from './Layout';
+export { Container } from './Container';
+export { Button } from './Button';
+export { Input } from './Input';
+export { SEO } from './SEO';
+export { Divider } from './Divider';
+export { SectionHeader } from './SectionHeader';
