+++ src/components/landing/About/index.js
@@ -1,0 +1,1 @@
+export { About } from './About';
