+++ src/components/landing/Contact/index.js
@@ -1,0 +1,1 @@
+export { Contact } from './Contact';
