+++ src/components/landing/Experience/index.js
@@ -1,0 +1,1 @@
+export { Experience } from './Experience';
