+++ src/components/landing/Intro/index.js
@@ -1,0 +1,1 @@
+export { Intro } from './Intro';
