+++ src/components/landing/Projects/index.js
@@ -1,0 +1,1 @@
+export { Projects } from './Projects';
