+++ src/components/landing/Projects/styles.js
@@ -12,7 +12,7 @@ export const Wrapper = styled.section`
 
 export const Grid = styled.div`
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(23rem, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(20rem, 1fr));
   grid-gap: 1.5rem;
 `;
