--- src/components/landing/Skills/index.jsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import AnchorLink from 'react-anchor-link-smooth-scroll';
-import { Container, Button } from 'components/common';
-import { Wrapper, SkillsWrapper, Details, Thumbnail } from './styles';
-
-export const Skills = () => (
-  <Wrapper id="about">
-    <SkillsWrapper as={Container}>
-      <Thumbnail>{/* <img src={dev} alt="I’m John and I’m a Backend & Devops engineer!" /> */}</Thumbnail>
-      <Details>
-        <h1>Hi There!</h1>
-        <p>
-          Lorem Ipsum is simply dummy text of the printing and typesetting
-          industry. Lorem Ipsum has been the industry’s standard dummy.
-        </p>
-        <Button as={AnchorLink} href="#contact">
-          Hire me
-        </Button>
-      </Details>
-    </SkillsWrapper>
-  </Wrapper>
-);
