+++ src/components/theme/Header/Hamburger/Hamburger.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { Wrapper, Bar } from './styles';
 
-const Hamburger = ({ sidebar, toggle }) => (
+export const Hamburger = ({ sidebar, toggle }) => (
   <Wrapper sidebar={sidebar} onClick={() => toggle(!sidebar)}>
     <Bar top sidebar={sidebar} />
     <Bar mid sidebar={sidebar} />
@@ -9,4 +10,7 @@ const Hamburger = ({ sidebar, toggle }) => (
   </Wrapper>
 );
 
-export default Hamburger;
+Hamburger.propTypes = {
+  sidebar: PropTypes.bool,
+  toggle: PropTypes.func,
+};
