+++ src/components/theme/Header/Hamburger/index.js
@@ -1,0 +1,1 @@
+export { Hamburger } from './Hamburger';
