+++ src/components/theme/Header/NavbarLinks/NavbarLinks.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import AnchorLink from 'react-anchor-link-smooth-scroll';
 import { Wrapper } from './styles';
 
-const NavbarLinks = ({ desktop }) => (
+export const NavbarLinks = ({ desktop }) => (
   <Wrapper desktop={desktop}>
     <AnchorLink href="#about">About</AnchorLink>
     <AnchorLink href="#experience">Experience</AnchorLink>
@@ -14,5 +14,3 @@ const NavbarLinks = ({ desktop }) => (
 NavbarLinks.propTypes = {
   desktop: PropTypes.bool,
 };
-
-export default NavbarLinks;
