+++ src/components/theme/Header/NavbarLinks/index.js
@@ -1,0 +1,1 @@
+export { NavbarLinks } from './NavbarLinks';
