+++ src/components/theme/Header/Sidebar/Sidebar.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { NavbarLinks } from '../NavbarLinks';
+import { Wrapper } from './styles';
+
+export const Sidebar = ({ sidebar, toggle }) => (
+  <Wrapper active={sidebar} onClick={toggle}>
+    <NavbarLinks />
+  </Wrapper>
+);
+
+Sidebar.propTypes = {
+  sidebar: PropTypes.bool,
+  toggle: PropTypes.func,
+};
