+++ src/components/theme/Tabs/index.js
@@ -1,0 +1,1 @@
+export { Tabs } from './Tabs';
